cluster:implemented code review, workers are made active or inactve in the user space by exposing the cluster scheduling handles
