+++ lib/cluster.js
@@ -522,6 +522,10 @@ function masterInit() {
       if (errno) delete handles[key];  // Gives other workers a chance to retry.
       send(worker, reply, handle);
     });
+    if (cluster.settings.shouldExposeSchedulingHandles) {
+      cluster.handles = cluster.handles || handles;
+      cluster.handles[key] = handle;
+    }
   }
 
   function listening(worker, message) {
